refactor : 닉네임 생성 로직 인터페이스로 분리
